refactor: move action into separate file to make testable
