+++ docs/api-reference/cli-options.mdx
@@ -58,7 +58,7 @@ Default: `mainnet`
 
 The directory to output the completed transfers. This directory will be used for both the completed transfers and any failed transfers.
 
-Completed transfers will be in a file named `{YYYY-MM-DD-HH-MM-SS}-{accounts-file-name}-completed.json`, and any failed transfers will be in a file named `{YYYY-MM-DD-HH-MM-SS}-{accounts-file-name}-failed.json`.
+Completed transfers will be in a file named `{timestamp}-{accounts-file-name}-completed.json`, and any failed transfers will be in a file named `{YYYY-MM-DD-HH-MM-SS}-{accounts-file-name}-failed.json`.
 
 Default: `~/.jumpdex`
