+++ src/enums/ErrorCodeEnum.ts
@@ -1,0 +1,6 @@
+enum ErrorCodeEnum {
+  Success,
+  Fail,
+}
+
+export default ErrorCodeEnum;
